build(api): make attach-addon func static as module export
Remove it from IUnitadeTabBuilder and make
correspondive file-modulae.
