+++ source/settings/tabs/tab-generic.ts
@@ -3,6 +3,7 @@ import UnitadeTabGenericBuilder from "@settings/tabs/factory/builder-generic";
 import { IUnitadeTab } from "@settings/tabs/tab";
 import { Addons } from "@settings/utils/types/addons";
 import UnitadePlugin from "@main";
+import attachAddon from "@settings/tabs/factory/ui/func/attach-addon";
 
 export default class UnitadeTabGeneric implements IUnitadeTab {
     tabContainer!: HTMLElement;
@@ -26,9 +27,10 @@ export default class UnitadeTabGeneric implements IUnitadeTab {
             this.tabBuilder.SETTING_CASE_INSENSITIVE
         ];
 
+        this.tabBuilder.updateState();
         this.tabBuilder.updateDisplays();
-        this.tabBuilder.attachAddon(this.tabBuilder.SETTING_CONFIG_EXTENSIONS_DEFAULT, Addons.Warning, 'EXTENSIONS WARNING');
 
+        attachAddon(this.tabBuilder.SETTING_CASE_INSENSITIVE, Addons.Warning, 'EXTENSIONS WARNING');
     }
 
     addEventListener(type: keyof HTMLElementEventMap, listener: (this: HTMLElement, ev: Event) => any, options?: boolean | AddEventListenerOptions): void {
